vote api

+++ app/router/post.py
@@ -20,11 +20,13 @@ router = APIRouter(
 )
 
 @router.get("/posts",response_model=List[schemas.Post])
-def get_posts(db: Session = Depends(get_db)
-              ,
-                user:int=Depends(authUtil.get_current_user)
+def get_posts(db: Session = Depends(get_db),
+                user:int=Depends(authUtil.get_current_user),
+                 limit: int = 10,
+                 skip: int = 0,
+                 search: str = ""
               ):
-    posts = db.query(models.Post).all()
+    posts = db.query(models.Post).filter(models.Post.title.contains(search)).limit(limit).offset(skip).all()
     return posts
 
 
@@ -32,9 +34,9 @@ def get_posts(db: Session = Depends(get_db)
 
 
 @router.get("/posts/{id}",response_model=schemas.Post)
-def get_post(id: int, db: Session = Depends(get_db)
-             ,
-                user:int=Depends(authUtil.get_current_user)
+def get_post(id: int, db: Session = Depends(get_db),
+             user:int=Depends(authUtil.get_current_user),
+            
              ):
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
@@ -75,8 +77,12 @@ def update_post(id: int, post: PostCreate, db: Session = Depends(get_db)
                 ):
     post_query = db.query(models.Post).filter(models.Post.id == id)
     updated_post = post_query.first()
+
     if not updated_post:
         raise HTTPException(status_code=404, detail="Post not found")
+    
+    if updated_post.owner_id != user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to perform requested action")
     post_query.update(post.dict(), synchronize_session=False)
     db.commit()
     return   post_query.first()
@@ -91,9 +97,14 @@ def delete_post(id: int, db: Session = Depends(get_db)
                 user:int=Depends(authUtil.get_current_user)
                 ):
     post_query = db.query(models.Post).filter(models.Post.id == id)
+
     deleted_post = post_query.first()
+   
     if deleted_post is None:
         raise HTTPException(status_code=404, detail="Post not found")
+    
+    if deleted_post.owner_id != user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to perform requested action")
     post_query.delete(synchronize_session=False)
     db.commit()
     return {"message": "Post deleted successfully"}
